hw1/provided-example/requirements.txt
crewai>=0.1.0,<0.28.0
crewai[tools]>=0.1.0,<0.28.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0